moved validation method to validation service class
