+++ src/main/java/uk/gov/hmcts/reform/bulkscan/services/BulkScanC100ValidationService.java
@@ -7,17 +7,31 @@ import uk.gov.hmcts.reform.bulkscan.model.BulkScanValidationResponse;
 import uk.gov.hmcts.reform.bulkscan.model.OcrDataField;
 import uk.gov.hmcts.reform.bulkscan.model.Status;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.ASKING_PERMISSION_FOR_APPLICATION;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.ATTENDED_MIAM;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILDREN_OF_SAME_PARENT;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILDREN_PARENTS_NAME;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILDREN_PARENTS_NAME_COLLECTION;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILDREN_SOCIAL_AUTHORITY;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILD_LIVING_WITH_APPLICANT;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILD_LIVING_WITH_OTHERS;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILD_LIVING_WITH_RESPONDENT;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILD_LOCAL_AUTHORITY_OR_SOCIAL_WORKER;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.EXEMPTION_TO_ATTEND_MIAM;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.EXISTING_CASE_ON_EMERGENCY_PROTECTION_CARE_OR_SUPERVISION_ORDER;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.FAMILY_MEMBER_INTIMATION_ON_NO_MIAM;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.MANDATORY_ATTENDED_MIAM_MESSAGE;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.MANDATORY_ERROR_MESSAGE;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.NO;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.PERMISSION_REQUIRED_REASON;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.PREVIOUS_OR_ONGOING_PROCEEDING;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.XOR_CONDITIONAL_FIELDS_MESSAGE;
+import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.YES;
 
 
 /**
@@ -113,6 +127,54 @@ public class BulkScanC100ValidationService {
 
             bulkScanValidationResponse.setStatus(Status.ERRORS);
         }
+    }
+
+    /**
+     * if Asking permission for application is selected as yes then permission reqiured reson.
+     * in section 5 is mandatory.
+     * */
+    public List<String> doPermissionRelatedFieldValidation(Map<String, String> inputFieldsMap) {
+        List<String> errors = new ArrayList<>();
+        if (null != inputFieldsMap.get(ASKING_PERMISSION_FOR_APPLICATION)
+            && YES.equalsIgnoreCase(inputFieldsMap.get(ASKING_PERMISSION_FOR_APPLICATION))
+            && null == inputFieldsMap.get(PERMISSION_REQUIRED_REASON)) {
+            errors.add(String.format(MANDATORY_ERROR_MESSAGE, PERMISSION_REQUIRED_REASON));
+        }
+        return errors;
+    }
+
+    /**
+     * 1. Checking if any one child living with option should be available
+     * 2. if Child same parent is yes then parent name should have value
+     * 3. if Child same parent is no then parent collection name should have value
+     * 4. if social authority is yes then local authority or social worker field should have value
+     * */
+    List<String> doChildRelatedValidation(Map<String, String> inputFieldsMap) {
+        List<String> errors = new ArrayList<>();
+        if (org.apache.commons.lang3.StringUtils.isEmpty(inputFieldsMap.get(CHILD_LIVING_WITH_APPLICANT))
+            && org.apache.commons.lang3.StringUtils.isEmpty(inputFieldsMap.get(CHILD_LIVING_WITH_RESPONDENT))
+            && org.apache.commons.lang3.StringUtils.isEmpty(inputFieldsMap.get(CHILD_LIVING_WITH_OTHERS))) {
+            errors.add(String.format(XOR_CONDITIONAL_FIELDS_MESSAGE, CHILD_LIVING_WITH_APPLICANT
+                .concat(",").concat(CHILD_LIVING_WITH_RESPONDENT).concat(",").concat(CHILD_LIVING_WITH_OTHERS)));
+        }
+
+        if (YES.equalsIgnoreCase(inputFieldsMap.get(CHILDREN_OF_SAME_PARENT))
+            && org.apache.commons.lang3.StringUtils.isEmpty(inputFieldsMap.get(CHILDREN_PARENTS_NAME))) {
+            errors.add(String.format(MANDATORY_ERROR_MESSAGE, CHILDREN_PARENTS_NAME));
+        }
+
+        if (NO.equalsIgnoreCase(inputFieldsMap.get(CHILDREN_OF_SAME_PARENT))
+            && org.apache.commons.lang3.StringUtils.isEmpty(inputFieldsMap.get(CHILDREN_PARENTS_NAME_COLLECTION))) {
+            errors.add(String.format(MANDATORY_ERROR_MESSAGE, CHILDREN_PARENTS_NAME_COLLECTION));
+        }
+
+        if (YES.equalsIgnoreCase(inputFieldsMap.get(CHILDREN_SOCIAL_AUTHORITY))
+            && org.apache.commons.lang3.StringUtils
+            .isEmpty(inputFieldsMap.get(CHILD_LOCAL_AUTHORITY_OR_SOCIAL_WORKER))) {
+            errors.add(String.format(MANDATORY_ERROR_MESSAGE, CHILD_LOCAL_AUTHORITY_OR_SOCIAL_WORKER));
+        }
+
+        return errors;
     }
 
     private Map<String, String> getOcrDataFieldsMap(List<OcrDataField> ocrDataFields) {
